ReplicaImp,preprocessor: important modification for transaction signing
After running Apollo tests, and due to new info acquired, I had to add
following changes:
1) All former metrics for transaction signing are removed from
pre-processor and replicaImp, and are no inside SigManager.
2) Exception ClientSignatureVerificationFailedException is removed, it
is unneeded. runtime_error exception is thrown when verification fail.
3) ClientRequestMsg message and signature validation is re-written.
4) Remove histogram from SigManager - verification is cheap, this is
unneeded.
5) PrePrepare message validates each message received, and when needed -
also, signatures are verified.
6) ReplicImp ctor - change init order of replicainfo, SigManager and
viewsManager
7) preprocessor - we first thought we can verify only in ReplicaImp -
this assumption was wrong. ClientPreProcessRequestMsg and
ClientBatchRequestMsg are verified on primary
(or any replica if the primary is unknown). It is then passed the
pre-processing engine, and also sent to other replicas using
PreProcessRequestMsg. the non-primary replica which received
PreProcessRequestMsg with the source as external client id, validates
the signature.
We also try our best to avoid validating the same request more than
once